Move functions for sig to slice
