+++ langwatch/src/components/evaluations/wizard/hooks/evaluation-wizard-store/slices/utils/nodeUtils.ts
@@ -8,3 +8,36 @@ export const calculateNodePosition = (lastNode: Node<Component> | undefined) =>
         y: lastNode.position.y,
       }
     : { x: 0, y: 0 };
+
+/**
+ * Generic parameter update function
+ * Will add the parameter if it doesn't exist and update the value if it does
+ */
+export function updateNodeParameter(
+  node: Node<Component>,
+  parameter: Partial<Omit<Field, "value">> & {
+    identifier: string;
+    type: Field["type"];
+    value?: any;
+  }
+): Node<Component> {
+  const parameters = node.data.parameters ?? [];
+  const paramIndex = parameters.findIndex(
+    (p) => p.identifier === parameter.identifier
+  );
+
+  const updatedParameters =
+    paramIndex === -1
+      ? [...parameters, parameter]
+      : parameters.map((param, index) =>
+          index === paramIndex ? { ...param, ...parameter } : param
+        );
+
+  return {
+    ...node,
+    data: {
+      ...node.data,
+      parameters: updatedParameters,
+    },
+  };
+}
